requirements.txt
asgiref==3.8.1
awsebcli==3.25
bcrypt==4.3.0
blessed==1.21.0
botocore==1.39.3
cement==2.10.14
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
decorator==5.2.1
Deprecated==1.2.18
Django==5.2.1
django-cors-headers==4.7.0
dj-database-url==2.1.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fabric==3.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
invoke==2.2.0
jmespath==1.0.1
packaging==24.2
paramiko==3.5.1
pathspec==0.12.1
pluggy==1.6.0
psycopg==3.2.9
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.4
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.13.2
urllib3==1.26.20
wcwidth==0.2.13
wrapt==1.17.2

